backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
python-dotenv==1.0.1
httpx==0.28.1
apscheduler==3.10.4
jinja2==3.1.5
aiosmtplib==3.0.2
pydantic-settings==2.7.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
